requirements.txt
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==8.2.0
Pillow==9.5.0
PyYAML==5.4.1
PyYAML==6.0
scikit_learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
torch==1.13.1
tqdm==4.59.0
